move files_versions related steps
